review fix ConventionForm
